Fix enabled/disable user action
As we do not set pwdMaxFailure at all, login failures do not count to
enter the disabled state: is should be safe to revert the value as in
version 1.0.1 (commit 9424ae60aa76fddb7354f75aff97090fe0046984).

Only the administrator can lock an account from the cluster-admin UI:
the commit fixes the behavior to this rule.